requirements.txt
click==6.7
python-telegram-bot==9.0.0
PyInstaller==3.3.1
